requirements.txt
python-telegram-bot==21.6
aiohttp
gspread
oauth2client
python-dotenv
